修改uint32_t与int32_t相等判断编译告警

+++ window_scene/session_manager/src/scene_session_manager.cpp
@@ -1847,7 +1847,7 @@ void SceneSessionManager::RecoverWindowSessionProperty(
         }
     } else {
         auto persistentId = sceneSession->GetPersistentId();
-        if (persistentId == callingWindowId_) {
+        if (persistentId == static_cast<int32_t>(callingWindowId_)) {
             WLOGFI("[WMSRecover] NotifyOccupiedAreaChangeInfo after calling session recovered,"
                 "persistentId = %{public}" PRId32, persistentId);
             sptr<OccupiedAreaChangeInfo> info = new OccupiedAreaChangeInfo();
